+++ sdks/go/pkg/beam/core/graph/coder/int.go
@@ -19,7 +19,7 @@ import (
 	"encoding/binary"
 	"io"
 
-	"github.com/apache/beam/sdks/go/pkg/beam/core/util/ioutilx"
+	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/util/ioutilx"
 )
 
 // EncodeUint64 encodes an uint64 in big endian format.